refactor: rename api function
