+++ src/api/getCharacters.ts
